moved safefile out of internal
